+++ examples/fsm/gensig2/main.ml
@@ -1,5 +1,4 @@
-(* FSM model of a signal generator.
-   Output [s] goes to 1 with input [start] and stays to 1 for [n] "cycles" *)
+(* This is a reformulation of ../gensig1 in which the output [s] is assigned to states *)
    
 open Utils
 open Lascar
@@ -19,12 +18,12 @@ let mk n = F.create
   ~inps:["start",[0;1]]
   ~outps:["s",[0;1]]
   ~vars:["k", ListExt.range Fun.id 0 n]
-  ~states:[Off,[]; On,[]]
+  ~states:[Off,["s",0]; On,["s",1]]
   ~istate:("", Off)
   ~trans:[
-    Off, ("start=1","k:=0;s:=1"), On;
+    Off, ("start=1","k:=0"), On;
     On, ("k<"^string_of_int n,"k:=k+1"), On;
-    On, ("k="^string_of_int n,"s:=0"), Off
+    On, ("k="^string_of_int n,""), Off
     ]
 
 let m1 = mk 2
@@ -33,8 +32,7 @@ let _ = F.dot_output "m1"  m1
 
 module FF = Conv.Fsm(F)
 
-let m2 =
-  FF.defactorize ~init:(Some ("",(Off,["k",0]))) [] m1
+let m2 = FF.defactorize ~init:(Some ("",(Off,["k",0]))) [] m1
 
 let _ = FF.dot_output ~options:[Dot.RankdirLR] "m2" m2