+++ CHANGES.md
@@ -4,6 +4,8 @@
     * fixed formatting bugs in generated doc 
     * Added functions `Ltsa.map_state_attr`, `Ltsa.map_transition`, `Lts.map_state`
       and `Lts.map_transition` and updated `examples/lts/map1` accordingly
+    * fixed bug in defactorization algorithm for FSMs with output values attached to
+      states (see for ex `examples/fsm/gensig2`)
 
 ## 0.6.0 (Sep, 16, 2019)
     * minor rewritings for ocaml >= 4.08